English doc RSTification part 1: The advancedtopics/ subdirectory
The image resources for the html files have also been moved to not
be on a higher tree level where images for all subdirs of the
English docs are stored. Now the ones for a particular subdir
reside in the pictures/ directory inside of that subdir.

The reason for this is that restview has issues with absolute paths,
and while Sphinx might be able to handle it, I don't know if
rst2html does. So err on the side of caution and put the relevant
images beneath the directory where they get called from.